Split tmux conversation pane
